# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "mfa_required": {
            "default": false,
            "title": "Mfa Required",
            "type": "boolean"
          },
          "mfa_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mfa Token"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "verification_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Url"
          }
        },
        "required": [
          "access_token",
          "user"
        ],
        "title": "AuthResponse",
        "type": "object"
      },
...
      "DemoPreview": {
        "description": "Расчёт портрета БЕЗ записи в базу.\n\n🔴 Схема заведена ДО фронта (v8.47.0): эндпоинт был размечен `-> dict[str, Any]`.\nЦена рукописного типа здесь высокая — предпросмотр снимает необратимость выбора\nпортрета (`/demo/load` стирает данные гостя), и расхождение вернуло бы человека\nк выбору вслепую.\n\n`metrics`, `plan` и `forecast` намеренно оставлены свободными объектами: их форму\nзадаёт движок (`run_planning`, `forecast_indicators`), у неё уже есть собственные\nсхемы в `/planning/*`, и дублировать их здесь значило бы завести вторую копию,\nрасходящуюся при первой же правке модели. Обязательность самих полей проверяется —\nименно на них строится экран.",
        "properties": {
          "expenses": {
            "$ref": "#/components/schemas/DemoMoneySide"
          },
          "forecast": {
            "additionalProperties": true,
            "title": "Forecast",
            "type": "object"
          },
          "goals": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Goals",
            "type": "array"
          },
          "history": {
            "additionalProperties": true,
            "title": "History",
            "type": "object"
          },
          "income": {
            "$ref": "#/components/schemas/DemoMoneySide"
          },
          "liquid_assets": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Liquid Assets",
            "type": "array"
          },
          "metrics": {
            "additionalProperties": true,
            "title": "Metrics",
            "type": "object"
          },
          "obligations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Obligations",
            "type": "array"
          },
          "plan": {
            "additionalProperties": true,
            "title": "Plan",
            "type": "object"
          }
        },
        "required": [
          "income",
          "expenses",
          "obligations",
          "goals",
          "liquid_assets",
          "history",
          "metrics",
          "plan",
          "forecast"
        ],
        "title": "DemoPreview",
        "type": "object"
      },
...
      "ForecastRequest": {
        "properties": {
          "horizon": {
            "default": 6,
            "maximum": 24.0,
            "minimum": 1.0,
            "title": "Horizon",
            "type": "integer"
          },
          "r_bench": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Сценарий «что если»: ставка капитализации вместо реальной OCR",
            "title": "R Bench"
          }
        },
        "title": "ForecastRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LegalDocumentContent": {
        "description": "Документ вместе с ТЕКСТОМ.\n\nОтдельно от `LegalDocument`: реестр отдаёт метаданные всех документов сразу и висит\nв футере каждой страницы — вкладывать в него шесть полных текстов значило бы возить\nдесятки килобайт на каждой навигации ради ссылки из четырёх слов.\n\nСодержимое — markdown, как в официальном пакете (`docs/legal/`), а не HTML. Конверсия\nна бэкенде завела бы ТРЕТЬЮ редакцию документа рядом с `md` и `docx`, и доказывать\nв споре пришлось бы, какая из трёх показана человеку.",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "effective_from": {
            "title": "Effective From",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "title",
          "version",
          "effective_from",
          "url",
          "slug",
          "content"
        ],
        "title": "LegalDocumentContent",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "PlanningCalculateResponse": {
        "description": "Ответ POST /planning/calculate.",
        "properties": {
          "admissible_count": {
            "title": "Admissible Count",
            "type": "integer"
          },
          "alternatives_total": {
            "title": "Alternatives Total",
            "type": "integer"
          },
          "best": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Alternative"
              },
              {
                "type": "null"
              }
            ]
          },
          "bliq_preallocation": {
            "$ref": "#/components/schemas/BliqPreallocation"
          },
          "crisis_plan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crisis Plan"
          },
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          },
          "indicators": {
            "$ref": "#/components/schemas/PlanningIndicators"
          },
          "input_summary": {
            "$ref": "#/components/schemas/InputSummary"
          },
          "ranked": {
            "items": {
              "$ref": "#/components/schemas/Alternative"
            },
            "title": "Ranked",
            "type": "array"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/Alternative"
            },
            "title": "Rejected",
            "type": "array"
          },
          "rejected_count": {
            "title": "Rejected Count",
            "type": "integer"
          },
          "risk_profile": {
            "title": "Risk Profile",
            "type": "string"
          },
          "surplus_plan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surplus Plan"
          },
          "top3": {
            "items": {
              "$ref": "#/components/schemas/Alternative"
            },
            "title": "Top3",
            "type": "array"
          },
          "weights": {
            "$ref": "#/components/schemas/Weights"
          }
        },
        "required": [
          "indicators",
          "bliq_preallocation",
          "risk_profile",
          "weights",
          "alternatives_total",
          "admissible_count",
          "rejected_count",
          "input_summary",
          "disclaimer"
        ],
        "title": "PlanningCalculateResponse",
        "type": "object"
      },
...
      "PlanningRequest": {
        "properties": {
          "expense_override": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Сценарий: расходы вместо фактических",
            "title": "Expense Override"
          },
          "income_override": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Сценарий: доход вместо фактического",
            "title": "Income Override"
          },
          "l_min": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Lmin — мин. допустимая Lt'",
            "title": "L Min"
          },
          "r_bench": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "OCR — порог Avalanche",
            "title": "R Bench"
          },
          "risk_tolerance": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Профиль риска (1–5)",
            "title": "Risk Tolerance"
          }
        },
        "title": "PlanningRequest",
        "type": "object"
      },
...
      "RecommendationRequest": {
        "properties": {
          "goals": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Goals",
            "type": "array"
          },
          "liquid_assets": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Liquid Assets",
            "type": "array"
          },
          "obligations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Obligations",
            "type": "array"
          },
          "transactions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "title": "RecommendationRequest",
        "type": "object"
      },
...
      "RegisterRequest": {
        "properties": {
          "consent": {
            "description": "Согласие на обработку ПДн (152-ФЗ). ОБЯЗАТЕЛЬНОЕ поле без значения по умолчанию: default=True на уровне контракта — это предотмеченная галочка, надзор читает её как навязанное согласие (юрблок L2). Явный false → отказ в регистрации.",
            "title": "Consent",
            "type": "boolean"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "newsletter_opt_in": {
            "default": false,
            "description": "Согласие на рекламную рассылку. Независимо от согласия на обработку ПДн: склеивать обязательное с необязательным нельзя.",
            "title": "Newsletter Opt In",
            "type": "boolean"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "referral_code": {
            "anyOf": [
              {
                "maxLength": 12,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Реферальный код пригласившего (необязательно).",
            "title": "Referral Code"
          }
        },
        "required": [
          "email",
          "password",
          "consent"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FINPILOT",
    "version": "8.46.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Вход",
        "tags": [
          "Аутентификация"
        ]
      }
    },
...
    "/api/auth/register": {
      "post": {
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Регистрация нового пользователя",
        "tags": [
          "Аутентификация"
        ]
      }
    },
...
    "/api/demo/preview": {
      "get": {
        "description": "Полный расчёт кейса без записи в БД: портрет, метрики, прогноз и рекомендация.\nДля раздела «Валидация» — визуальная проверка всех основных функций сразу.",
        "operationId": "preview_demo_api_demo_preview_get",
        "parameters": [
          {
            "in": "query",
            "name": "case",
            "required": false,
            "schema": {
              "default": "anna",
              "title": "Case",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Портрет кейса + метрики + прогноз + рекомендация",
        "tags": [
          "Демо-данные"
        ]
      }
    },
...
    "/api/legal/documents/{slug}": {
      "get": {
        "description": "Текст документа из ЕДИНСТВЕННОГО источника — файла официального пакета.\n\n🔴 Публичный намеренно. Политику обработки ПДн человек обязан прочитать ДО\nрегистрации, иначе согласие неинформированное — а это дефект юридический, а не\nинтерфейсный. Требовать авторизации, чтобы прочитать условия обработки данных,\nзначит замкнуть круг.\n\n🔴 Путь берётся из реестра ПО КЛЮЧУ и никогда не склеивается из ввода: иначе\n`../../.env` читался бы с диска через публичный эндпоинт. Неизвестный ключ — 404\nдо всякого обращения к файловой системе.",
        "operationId": "legal_document_content_api_legal_documents__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalDocumentContent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Текст юридического документа",
        "tags": [
          "Согласия и право"
        ]
      }
    },
...
    "/api/planning/calculate": {
      "post": {
        "operationId": "calculate_plan_api_planning_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanningRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningCalculateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Полный цикл СППР: генерация и ранжирование альтернатив",
        "tags": [
          "Планирование"
        ]
      }
    },
...
    "/api/planning/forecast": {
      "post": {
        "operationId": "get_forecast_api_planning_forecast_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Forecast Api Planning Forecast Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Прогноз Rt/Lt/Dt на горизонт H (форм. 35 ВКР)",
        "tags": [
          "Планирование"
        ]
      }
    },
...
    "/api/recommendation": {
      "post": {
        "operationId": "create_recommendation_api_recommendation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RecommendationRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Быстрая текстовая рекомендация по показателям финансового состояния",
        "tags": [
          "Рекомендации"
        ]
      }
    },
...
    "/api/transactions": {
      "get": {
        "operationId": "get_transactions_endpoint_api_transactions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "title": "Response Get Transactions Endpoint Api Transactions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Получить список транзакций",
        "tags": [
          "Транзакции"
        ]
      },
      "post": {
        "operationId": "create_transaction_endpoint_api_transactions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Создать транзакцию",
        "tags": [
          "Транзакции"
        ]
      }
    },
...
    "/api/transactions/{transaction_id}": {
      "delete": {
        "operationId": "delete_transaction_endpoint_api_transactions__transaction_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "title": "Transaction Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Удалить транзакцию",
        "tags": [
          "Транзакции"
        ]
      },
      "put": {
        "operationId": "update_transaction_endpoint_api_transactions__transaction_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "title": "Transaction Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Изменить операцию (частично, только переданные поля)",
        "tags": [
          "Транзакции"
        ]
      }
    },
...
    "/v1/analyze": {
      "post": {
        "operationId": "analyze_v1_analyze_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "B2B: анализ канонического снимка финансов партнёра",
        "tags": [
          "B2B"
        ]
      }
    },
...
  }
}